add some refactor
