+++ src/SeaIceThermodynamics/SlabThermodynamics/slab_tendency_kernel_functions.jl
